### 2023-01-27 #5107 @editorialbot
[PRE REVIEW]: dsmmR: Estimation and Simulation of Drifting Semi-Markov Models
**Submitting author:** @Mavrogiannis-Ioannis (<a href="http://orcid.org/0000-0002-2948-0648">Ioannis Mavrogiannis</a>)
**Repository:** https://github.com/Mavrogiannis-Ioannis/dsmmR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
