+++ paper.bib
@@ -3,8 +3,7 @@
   author={Barbu, Vlad Stefan and Limnios, Nikolaos},
   volume={191},
   year={2009},
-  publisher={Springer Science \& Business Media},
-  url = {https://doi.org/10.1007/978-0-387-73173-5}
+  publisher={Springer Science \& Business Media}
 }
 
 
@@ -13,15 +12,6 @@
   author={Barbu, Vlad Stefan and Lecocq, Florian and Lothode, Corentin}
 }
 
-
-@Article{smm_package,
-  title={SMM: An R Package for Estimation and Simulation of Discrete-time semi-{Markov} Models.},
-  author={Barbu, Vlad Stefan and Vergne, Nicolas},
-  journal={R Journal},
-  volume={10},
-  number={2},
-  year={2018}
-}
 
 
 @Manual{smmR_package,
@@ -34,13 +24,11 @@
 
 
 @Article{drift_polynomial,
-url = {https://doi.org/10.2202/1544-6115.1326},
 title = {Drifting {Markov} Models with Polynomial Drift and Applications to DNA Sequences},
 author = {Nicolas Vergne},
 volume = {7},
 number = {1},
 journal = {Statistical Applications in Genetics and Molecular Biology},
-doi = {doi:10.2202/1544-6115.1326},
 year = {2008},
 lastchecked = {2022-10-16}
 }
@@ -64,13 +52,6 @@ lastchecked = {2022-10-16}
     year = {2021},
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=drimmR},
-}
-
-
-@Article{drifting_swimming,
-  title={Learning the arm-leg coordination in breaststroke swimming: {Identifying} exploration of movement dynamics with drifting {Markov} {Models}},
-  author={Komar, John and Seifert, Ludovic and Vergne, Nicolas and Newell, Karl M},
-  year={2022}
 }
 
 
