+++ README.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 ```
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
+<!-- Use `build_readme()` to edit the .md file from the .rmd file -->
 <!-- badges: start -->
 
 [![CRAN
