+++ paper.bib
@@ -30,7 +30,8 @@ volume = {7},
 number = {1},
 journal = {Statistical Applications in Genetics and Molecular Biology},
 year = {2008},
-lastchecked = {2022-10-16}
+lastchecked = {2022-10-16},
+url = {https://doi.org/10.2202/1544-6115.1326}
 }
 
 
@@ -42,7 +43,8 @@ lastchecked = {2022-10-16}
   number={4},
   pages={1407--1429},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  url = {https://doi.org/10.1007/s11009-018-9682-8}
 }
 
 
