+++ paper.bib
@@ -24,14 +24,14 @@
 
 
 @Article{drift_polynomial,
-title = {Drifting {Markov} Models with Polynomial Drift and Applications to DNA Sequences},
-author = {Nicolas Vergne},
-volume = {7},
-number = {1},
-journal = {Statistical Applications in Genetics and Molecular Biology},
-year = {2008},
-lastchecked = {2022-10-16},
-url = {https://doi.org/10.2202/1544-6115.1326}
+  title = {Drifting {Markov} Models with Polynomial Drift and Applications to DNA Sequences},
+  author = {Nicolas Vergne},
+  volume = {7},
+  number = {1},
+  journal = {Statistical Applications in Genetics and Molecular Biology},
+  year = {2008},
+  lastchecked = {2022-10-16},
+  doi = {10.2202/1544-6115.1326}
 }
 
 
@@ -44,7 +44,7 @@ url = {https://doi.org/10.2202/1544-6115.1326}
   pages={1407--1429},
   year={2019},
   publisher={Springer},
-  url = {https://doi.org/10.1007/s11009-018-9682-8}
+  doi = {10.1007/s11009-018-9682-8}
 }
 
 
