### 2023-03-13 #5107 @Bisaloo
Hi @chjackson :wave:, would you be interested in reviewing this [R package](https://github.com/Mavrogiannis-Ioannis/dsmmR) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05107/joss.05107/10.21105.joss.05107.pdf) for the Journal of Open Source Software (JOSS)?

> dsmmR: Estimation and Simulation of Drifting Semi-Markov Models

I found you via your work on the msm R package and I believe your expertise would be valuable here.

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.

If you are not available to review this submission, do you have any recommendation for suitable reviewers?
