### 2023-03-13 #5107 @chjackson
@Bisaloo Sorry to decline this, I'm very busy at the moment.  Not sure that I know anyone with expertise particularly suited to this style of model, other than the authors of the packages cited in the paper.
