### 2023-03-28 #5107 @m-clark
Hi Hugo,

Thanks for reaching out and apologies for the belated reply, I was out on
vacation part of last week.  Unfortunately, I'm not able to offer a review
at this time, as I am not able to do this sort of thing during my normal
work schedule, leaving little extra time for that sort of thing. I also am
not familiar with these models, so probably would not be a good judge of
content without spending time getting more context.  In the end, I wish you
the best of luck and hope the package proves useful!

Michael

On Mon, Mar 20, 2023 at 6:39 AM Hugo Gruson ***@***.***>
wrote:

> Hi @m-clark <https://github.com/m-clark> 👋, would you be interested in
> reviewing this R package <https://github.com/Mavrogiannis-Ioannis/dsmmR>
> and the associated paper
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.05107/joss.05107/10.21105.joss.05107.pdf>
> for the Journal of Open Source Software (JOSS)?
>
> dsmmR: Estimation and Simulation of Drifting Semi-Markov Models
>
> Your very interesting 'Model Estimation by Example' book
> <https://m-clark.github.io/models-by-example/> made me think of you for
> this submission.
>
> You can have a look at our review criteria
> <https://joss.readthedocs.io/en/latest/review_criteria.html> and review
> checklist <https://joss.readthedocs.io/en/latest/review_checklist.html>
> to get an idea of what is expected.
>
> If you are not available to review this submission, do you have any
> recommendation for suitable reviewers?
>
