### 2023-04-24 #5107 @Mavrogiannis-Ioannis
Hello @Bisaloo, sorry for my absence. I hope you are doing well and that life has been kind to you.

If we could reach some reviewers like @ellessenne, @Athene-ai, @dsurujon, which are the reviewers of  the smmR package, it would be excellent.

Ioannis
