### 2023-04-24 #5107 @Athene-ai
I can review

On Mon, 24 Apr 2023 at 22:30, Ioannis Mavrogiannis ***@***.***>
wrote:

> Hello @Bisaloo <https://github.com/Bisaloo>, sorry for my absence. I hope
> you are doing well and that life has been kind to you.
>
> If we could reach some reviewers like @ellessenne
> <https://github.com/ellessenne>, @Athene-ai <https://github.com/Athene-ai>,
> @dsurujon <https://github.com/dsurujon>, which are the reviewers of the
> smmR package, it would be excellent.
>
> Ioannis
>
