### 2023-06-19 #5107 @Bisaloo
@RomainAzais has informed me that he contributed a chapter to a book edited by submission authors.

While this is not completely defined in our conflict of interest policies, I have decided to officially waive it since this is akin to contribute to a special edition in a journal, and does not create a relationship that we deem likely to impact the quality and honesty of the review.

Since we now have two reviewers for this submission, I will close this pre-review issue and create the actual review issue, where our conversations will take place from now on.
