### 2023-06-19 #5572 @Bisaloo
:wave: :wave: :wave: @Mavrogiannis-Ioannis @Athene-ai @RomainAzais this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 2-4 weeks. However, I have noted that @RomainAzais is not available to start the review before ~mid-July. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
