+++ DESCRIPTION
@@ -2,7 +2,7 @@ Type: Package
 Package: dsmmR
 Title: Estimation and Simulation of Drifting Semi-Markov Models
 Version: 1.0.2
-Date: 2023-02-03
+Date: 2023-08-31
 Authors@R: c(
     person("Vlad Stefan", "Barbu",
            role = c("aut"),
@@ -28,7 +28,12 @@ Description: Performs parametric and non-parametric estimation and simulation of
     Drifting Markov models are described in:
     Vergne, N. (2008) <doi:10.2202/1544-6115.1326>.
     Reliability indicators of Drifting Markov models are described in:
-    Barbu, V. S., Vergne, N. (2019) <doi:10.1007/s11009-018-9682-8>.
+    Barbu, V. S., Vergne, N. (2019) <doi:10.1007/s11009-018-9682-8>. 
+    We acknowledge the DATALAB Project
+    <https://lmrs-num.math.cnrs.fr/projet-datalab.html> (financed by the
+    European Union with the European Regional Development fund (ERDF) and by
+    the Normandy Region) and the HSMM-INCA Project (financed by the French
+    Agence Nationale de la Recherche (ANR) under grant ANR-21-CE40-0005).
 License: GPL
 Imports: DiscreteWeibull
 Suggests: 
