### 2023-09-01 #5572 @Mavrogiannis-Ioannis
Hello and good morning everyone,
Yesterday I made a small update to the github page of the package, specifically:
adding an Acknowledgements section in the README and DESCRIPTION files, regarding the financing that made the package possible.

It does not affect the paper or the code in any way.
In fact, the financing was already mentioned in the paper.md file that you are reviewing.
I hope you enjoyed your "vacations"!
Hoping to hear from you, soon.

Ioannis
