### 2023-10-10 #5572 @RomainAzais
Hello @Mavrogiannis-Ioannis

I have a few first questions about the package and, more specifically, about model simulation.

If I understand correctly, the package only allows you to simulate model 3 where only the sojourn times drift. Why this restriction?

Beyond that, the output of the simulation function is the sequence of states, without their sojourn times. Don't you think this limits the package's interest in simulation? I also find the documentation a little ambiguous on this point.

Thank you for your feedback on these points.
