### 2023-10-11 #5572 @Mavrogiannis-Ioannis
Hello @RomainAzais!

Thank you kindly for your feedback. I will do my best to understand your points and explain my point of view.

a. "If I understand correctly, the package only allows you to simulate model 3 where only the sojourn times drift. Why this restriction?"

The package allows the user to simulate a sequence from any given model. This can be model 1, 2 or 3.
This can be seen in the `Arguments` section of `simulate.dsmm()`,
 where `obj` inherits from all possible objects, namely
`dsmm` and one of:
`dsmm_fit_nonparametric`, `dsmm_nonparametric`, `dsmm_fit_parametric` or `dsmm_parametric`.

Since these objects can have any model (seen in earlier documentation of functions `fit_dsmm()`, `parametric_dsmm()` and `nonparametric_dsmm()`)
then we can have any model for the simulate function.

For example, you can run the `Examples` where you simply switch `f_is_drifting` and `p_is_drifting` to obtain Model 1 and 2.

Perhaps your initial concern lies in the `Examples` section of the function, where we only show an example for the 3rd model.

We chose this way of representation to showcase the functionality of `simulate.dsmm()`: you can give it any model you previously had
(in the example's case `obj_model_3`) and it's easy to simulate from that point.

b. "Beyond that, the output of the simulation function is the sequence of states, without their sojourn times. Don't you think this limits the package's interest in simulation?...

That is true, the output value is just the "regular" sequence of states. This way the researcher can do as they wish with the sequence.
They can obtain the embedded Markov chain and the corresponding sojourn times by running the function `base::rle()`.
From this point of view, it is a simple matter to add to the output of `simulate.dsmm()`, I can do so at the weekend. Do you consider it necessary?

c. "... I also find the documentation a little ambiguous on this point."

Can you specify exactly which part?

Again, thank you a lot for your feedback. Hope to hear from you soon,
Ioannis
