### 2023-10-16 #5572 @RomainAzais
Hello @Mavrogiannis-Ioannis

In fact, I hadn't quite understood. Thanks for your reply: it's much clearer now.

Adding to the documentation that `base::rle()` can be used to obtain the embedded chain and sojourn times is a good idea indeed.

One might also expect `simulate.dsmm()` to return an object of the S3 class (keeping the parameters used for simulation). I don't know if it's a good idea, but any thoughts on this?

Thank you again for your explanations.
Romain
