### 2023-10-19 #5572 @Mavrogiannis-Ioannis
Hello @RomainAzais , thank you again for your insightful comments.

It is indeed intriguing for a `dsmm_simulate` object to exist, inheriting parameters from the object given in the `simulate.dsmm()` function; this way it will be easier to track down different simulated sequences, making it easy to use for researchers.
Added documentation for the `base::rle()` function would indeed be useful for everyone using the software as well.

I will think about how to formulate this new object (so that it fits in to the other objects in the `dsmm` "family") and hopefully present it here in the week-end.

Automated tests will also be needed for this new object, similarly to all the other objects in the package. This will take perhaps longer to make, to make sure it's extensive. This will be done so that people can create their own `dsmm_simulate` object (although it's always easier to use the `simulate.dsmm()` function).

Thank you again for your feedback!
Ioannis
