### 2023-11-02 #5572 @Mavrogiannis-Ioannis
So, after a long consideration, I propose the following:

- I will add in detailed example usage for `base::rle()` function in the `#Examples` section of both `simulate.dsmm()` AND `fit_dsmm()`. This way, I believe it will be clear to the users about the difference between a regular sequence (example: AAACTGGT), the embedded Markov chain (for the example: ACTGT) and the corresponding sojourn times (for the example: 31121).
- Nothing else will change for the `simulate.dsmm()` function.
- An important (yet subtle) change will be made in the `fit_dsmm()` function, following the same direction: change the returned object's `seq` attribute to `emc`.

I believe this is a good way to achieve the most clarity, while keeping the objects simple, without repeating too much of the same information across objects and variables. The user should be able to keep track of the models and sequences they are using. They only have to properly name their own variables.

Please, let me know your thoughts on this approach.

Thank you for your time and your patience,
Ioannis
