### 2023-11-03 0d50800
Adding clarity for sequence and the EMC. Updating fit_dsmm and simulate.dsmm with these changes.
