+++ R/generic_functions.R
@@ -54,9 +54,9 @@ check_attributes.dsmm_fit_nonparametric <- function(obj) {
                     f_is_drifting = f_is_drifting
                     # , numerical_est = numerical_est
         ),
-        valid_seq(seq = (seq <- obj$seq)),
+        valid_seq(emc = (emc <- obj$emc)),
         valid_model_size(model_size = (model_size <- obj$model_size),
-                         length_seq = (l <- length(seq))),
+                         length_seq = (l <- length(emc))),
         valid_soj_times(soj_times = (soj_times <- obj$soj_times),
                         length_seq = l),
         valid_k_max(k_max = (k_max <- obj$k_max), soj_times = soj_times),
@@ -76,7 +76,7 @@ check_attributes.dsmm_fit_nonparametric <- function(obj) {
         )
     )
     # Check names of the list.
-    names_fit <- c("dist", "seq", "soj_times", "initial_dist", "states",
+    names_fit <- c("dist", "emc", "soj_times", "initial_dist", "states",
                    "s", "degree", "k_max", "model_size", "f_is_drifting",
                    "p_is_drifting", "Model", "estimation", "A_i", "J_i")
     if (!identical((nobj <- names(obj)), names_fit)) {
@@ -135,9 +135,9 @@ check_attributes.dsmm_fit_parametric <- function(obj) {
                     f_is_drifting = f_is_drifting
                     # , numerical_est = numerical_est
         ),
-        valid_seq(seq = (seq <- obj$seq)),
+        valid_seq(emc = (emc <- obj$emc)),
         valid_model_size(model_size = (model_size <- obj$model_size),
-                         length_seq = (l <- length(seq))),
+                         length_seq = (l <- length(emc))),
         valid_soj_times(soj_times = (soj_times <- obj$soj_times),
                         length_seq = l),
         valid_k_max(k_max = (k_max <- obj$k_max), soj_times = soj_times),
@@ -157,7 +157,7 @@ check_attributes.dsmm_fit_parametric <- function(obj) {
         )
     )
     # Check names of the list.
-    names_fit <- c("dist", "seq", "soj_times", "initial_dist", "states",
+    names_fit <- c("dist", "emc", "soj_times", "initial_dist", "states",
                    "s", "degree", "k_max", "model_size", "f_is_drifting",
                    "p_is_drifting", "Model", "estimation", "A_i", "J_i")
     if (!identical((nobj <- names(obj)), names_fit)) {
@@ -464,8 +464,7 @@ is.dsmm_parametric <- function(obj) {
 #'
 #' @description
 #' This is a generic method that computes and returns the drifting
-#' semi-Markov kernel as a numerical array of dimensions
-#' \eqn{s \times s \times k_{max} \times (n + 1)}.
+#' semi-Markov kernel.
 #'
 #' @param obj An object that inherits from the S3
 #' classes \code{dsmm},
@@ -598,9 +597,9 @@ is.dsmm_parametric <- function(obj) {
 #' @examples
 #' # Setup.
 #' states <- c("Rouen", "Bucharest", "Samos", "Aigio", "Marseille")
-#' seq <- create_sequence(states, probs = c(0.3, 0.1, 0.1, 0.3, 0.2))
+#' emc <- create_sequence(states, probs = c(0.3, 0.1, 0.1, 0.3, 0.2))
 #' obj_model_2 <- fit_dsmm(
-#'     sequence = seq,
+#'     sequence = emc,
 #'     states = states,
 #'     degree = 3,
 #'     f_is_drifting = FALSE,
@@ -829,7 +828,7 @@ print.dsmm <- function(x, ...) {
     check_attributes(x)
     nm <- names(x)
     for (i in seq_along(nm)) {
[…]
