+++ R/fitmodel.R
@@ -11,7 +11,7 @@
 
 
 #' @title Estimation of a drifting semi-Markov chain
-#' @aliases dsmm_fit dsmm_fit_nonparametric dsmm_fit_parametric
+#' @aliases dsmm_fit
 #' @description Estimation of a drifting semi-Markov chain,
 #' given one sequence of states. This estimation can be parametric
 #' or non-parametric and is available for
@@ -341,8 +341,7 @@
 #' @return Returns an object of S3 class \code{(dsmm_fit_nonparametric, dsmm)}
 #' or \code{(dsmm_fit_parametric, dsmm)}. It has the following attributes:
 #' \itemize{
-#' \item \code{dist} : List. Contains 2 or 3 arrays,
-#' \code{estimation}:
+#' \item \code{dist} : List. Contains 2 or 3 arrays.
 #' \itemize{
 #' \item If \code{estimation = "nonparametric"} we have 2 arrays:
 #'   \itemize{
@@ -363,14 +362,14 @@
 #'     depending on whether \eqn{f} is drifting or not.
 #'   }
 #' }
-#' \item \code{seq} : Character vector that contains the
+#' \item \code{emc} : Character vector that contains the
 #'   \strong{embedded Markov chain} \eqn{(J_{t})_{t\in \{0,\dots,n\}}}
 #'   \strong{of the original sequence}.
 #'   It is this attribute of the object that describes the size of the model
 #'   \eqn{n}. Last state is also included, for a total length of \eqn{n+1},
 #'   but it is not used for any calculation.
 #' \item \code{soj_times} : Numerical vector that contains the sojourn times
-#'   spent for each state in \code{seq} before the jump to the next state.
+#'   spent for each state in \code{emc} before the jump to the next state.
 #'   Last state is also included, for a total length of \eqn{n+1},
 #'   but it is not used for any calculation.
 #' \item \code{initial_dist} : Numerical vector that contains an estimation
@@ -390,7 +389,7 @@
 #'   drifting semi-Markov model \eqn{n}, which is equal to the length of the
 #'   embedded Markov chain \eqn{(J_{t})_{t\in \{0,\dots,n\}}},
 #'   minus the last state.
-#'   It has a value of \code{length(seq) - 1}, for \code{seq} as defined above.
+#'   It has a value of \code{length(emc) - 1}, for \code{emc} as defined above.
 #' \item \code{f_is_drifting} : Logical. Passing down from the arguments.
 #' Specifies if \eqn{f} is drifting or not.
 #' \item \code{p_is_drifting} : Logical. Passing down from the arguments.
@@ -457,11 +456,11 @@
 #' states <- sort(unique(sequence))
 #' degree <- 3
 #'
+#'
 #' # ===========================================================================
 #' # Nonparametric Estimation.
 #' # Fitting a random sequence under distributions of unknown shape.
 #' # ===========================================================================
-#'
 #'
 #'
 #' # ---------------------------------------------------------------------------
@@ -493,6 +492,11 @@
 #'            "Dimension of f_drift: (s, s, k_max, d + 1) = (",
 #'            paste(dim(f_drift), collapse = ", "), ").\n"))
 #'
+#' # We can even check the embedded Markov chain and the sojourn times
+#' # directly from the returned object, if we wish to do so.
+#' # This is achieved through the `base::rle()` function, used on `sequence`.
+#' model_emc <- obj_model_1$emc
+#' model_sojourn_times <- obj_model_1$soj_times
 #'
 #' # ---------------------------------------------------------------------------
[…]

[…]
