+++ README.Rmd
@@ -24,6 +24,10 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
+# Developer Version
+
+## 1.0.3
+
 # dsmmR
 
 The **dsmmR** R package allows the user to estimate, simulate and define
