### 2023-11-04 #5572 @Mavrogiannis-Ioannis
Hello,
the changes as explained on my previous comment are live on github : https://github.com/Mavrogiannis-Ioannis/dsmmR/tree/master

Thank you everyone for your work so far.
I'm waiting for more of your comments.
Best,
Ioannis
