### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
Add automated tests
One important way to ensure your package is behaving as intended and producing the expected results is to add automated tests.

This is part of https://github.com/openjournals/joss-reviews/issues/5572 as JOSS policy states:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

Good resources to get started with automated tests in R packages:

- [The dedicated chapter in the 'R packages' book by Hadley Wickham and Jenny Bryan](https://r-pkgs.org/testing-basics.html)
- [The testthat package documentation](https://testthat.r-lib.org/index.html)

Please let me know if you'd like additional assistance or guidance to set this up.
