### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Mavrogiannis-Ioannis
Hello @Bisaloo. There are extended automated tests in place, that run when a function is used. You can see them mostly in the utils.R file. These are automatically used when the examples are run, when the package is being checked with Cran etc. So, if there is no error, they work perfectly.
Please Le me know if this is okay with you.
Best regards,
Ioannis
