### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
Hi, thanks for your answer. I'd recommend migrating them to the standard testing location at some point but I believe this is not a blocking issue for the JOSS publication.

Could you however increase the code coverage? It seems some portions of the code are not tested:

```r
covr::package_coverage(type = "examples")
```
> dsmmR Coverage: 61.12%
> R/generic_functions.R: 51.10%
> R/parametric_estimation.R: 55.00%
> R/utils.R: 57.60%
> R/nonparametric.R: 76.83%
> R/parametric.R: 79.07%
> R/fitmodel.R: 81.30%
