### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Mavrogiannis-Ioannis
Thank you for your response. Can you show me one example where the code fails to recognize a mistake, so that I know what to look for?
Thank you a lot,
Ioannis
