### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
I think we are talking about two different things.

From what I understand, you are talking about input checking: ensuring that the user gets a meaningful error message if they pass invalid parameter.

I'm talking about [automated tests](https://www.atlassian.com/continuous-delivery/software-testing/automated-testing) that automatically verify that the code is working as intended and that future updates to the code don't accidentally introduce a bug that could go undetected.

Or am I misunderstanding?
