### 2023-11-11 Mavrogiannis-Ioannis/dsmmR#2 @Mavrogiannis-Ioannis
Okay I understand. Yes, they are different things. However there is a small overlap, as I understand it. All the test cases (that a normal user would come across) seem to be included in the examples... So most of the bugs would show up there.sa

What you propose is of course the most robust way to go about it, but it will require a lot of coding. Do you think we have enough right now?

Also, which functions should we first make automated test for? Personally I think the fit function should be first, then the simulate function, then the get_kernel function, then the parametric non/ parametric and finally the utils.R file.

Thank you for your input and your constant support.
Ioannis
