### 2023-11-17 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
> Okay I understand. Yes, they are different things. However there is a small overlap, as I understand it. All the test cases (that a normal user would come across) seem to be included in the examples... So most of the bugs would show up there.sa
>
>What you propose is of course the most robust way to go about it, but it will require a lot of coding. Do you think we have enough right now?

Agreed on both points. Your current approach indirectly serves as tests and I don't think there is a need to switch right now. I do encourage you to do it at some point but this is not a blocker for JOSS.

> Also, which functions should we first make automated test for? Personally I think the fit function should be first, then the simulate function, then the get_kernel function, then the parametric non/ parametric and finally the utils.R file.

You should investigate which portions of the code are never reached by your examples. This can be done with the covr package. For example, the following snippet with list the lines that examples never execute:

```r
covr::package_coverage(type = "examples") |> covr::zero_coverage()
```
