### 2023-11-24 #5572 @RomainAzais
Hello @Mavrogiannis-Ioannis

Thanks for the food for thought and the changes.
I tested the new version and read the explanations behind it: I think it is clear and easy to use now.

Regarding the package's functionalities, I think it would be worth clarifying that estimation requires only one trajectory observation (in long time) and not several trajectories (both frameworks being allowed in statistics of processes).

On this aspect, the `nsim` input variable of the `simulate.dsmm()` function can be confusing (especially with the other optional variable `seq_length`): maybe prefer names like `horizon` and `max_horizon`? (These are really minor points.)
