### 2023-11-24 #5572 @Mavrogiannis-Ioannis
Dear @RomainAzais,

First of all, I just discussed with my one of my supervisors during the `dsmmR` package development, Nicolas Vergne, and we are going to add multiple trajectories as a possibility for the user as well. So I will make this part more well defined when I write the code (in 1-2 weekends...).

As for the `simulate.dsmm()`, I agree that it requires to spend some time to understand the `seq_length` attribute etc.

I want to ask you, are names like `horizon` and `max_horizon` well defined in the literature?
If so, is it possible for you to refer to me an example usage of these terms?

Thank you very much for your support & patience.

Best regards,
Ioannis
