### 2023-11-24 #5572 @RomainAzais
When simulating a stochastic process (X_t) over a time window [0,T], T is often referred to as the horizon. This may be particularly true in the context of stochastic algorithms or optimal control. See, for example, [these lecture notes](http://researchers.lille.inria.fr/~lazaric/Webpage/MVA-RL_Course14_files/notes-lecture-02.pdf) on Markov decision processes.

`max_horizon` could denote an upper-bound of the horizon: if I remember correctly, if `nsim` is greater than `seq_length`, the simulation stops at `seq_length`.

By the way, it is great to be able to use multiple trajectories for estimation!

Best,
Romain
