### 2023-12-17 1538445
In the function `simulate.dsmm`, changed the attribute `seq_length` to `max_seq_length` for clarification.
