### 2023-12-17 Mavrogiannis-Ioannis/dsmmR#2 @Mavrogiannis-Ioannis
Hello @Bisaloo , thank you again for this helpful comment.
I run the code, and it gave me more insight about what causes the coverage being so "low", and also which parts of the code the examples never reach.

- Every line of a stop statement that I put in the code is counted as "No Test Coverage!". I'm sure there is a better way that what I've done, but it seems pretty harmless.
- Others include `if` statements about some variables that are created inside functions, that check if something happens that shouldn't happen; e.g. the user intentionally gives a value when defining the models that is wrong, like a string instead of a number.
- Some more are about some obscure details that should never happen, and again consider the user intentionally trying to break the code, defining models with wrong names, etc.

So it makes sense that the above are not reached by the examples; the examples are well-constructed, and showcase the normal usage, not explicitly the extreme testing under the hood.

I will look into this further, but perhaps when I decide to implement automated tests.

Thank you again for giving insight into this aspect of the code!
Respectfully,
Ioannis
