### 2023-12-17 #5572 @Mavrogiannis-Ioannis
Hello @RomainAzais , thank you again for providing references for your previous comment, it is quite interesting!

After pondering about it, I think the best course of action is as follows, regarding the attributes `nsim` and `seq_length` of the `simulate.dsmm()` function:

- `nsim` : **Name stays the same**, since it also is "inherited" from the generic `simulate()` function used in base R and users should be familiar with it.
 - Also a comment for us to think about the possibility of the `horizon` and `max_horizon` naming for `nsim` and `seq_length`:
    *  For our stochastic model `(X_t)`, that acts behind the simulate function, we will have `t` ranging in `[0, T]` but notice here that `T` is the `model_size` that is mentioned before. Notice also that `nsim` has a maximum value of `model_size`. This means that here we are moving alongside the Embedded Markov Chain (EMC).
    * The resulting sequence from the `simulate.dsmm()` function, will be the combination of the EMC **plus** the Sojourn Times of each state; therefore, `seq_length` is moving alongside a different "scale" than `nsim`. So although it is interesting to propose `horizon` and `max_horizon`, they are moving alongside different time scales, perhaps it is not appropriate?
This is just for us to think about, and also I'm not sure if I understand clearly the horizon/max horizon cases. So perhaps we can stick to the simple case.

- `seq_length` : **Name changes** to `max_seq_length`. I think this name is easier to understand, as one can literally translate it to `maximum sequence length`, i.e. a limiting value for the length of the resulting sequence (both the EMC and the Soj. Times).

I am going to work on the multiple sequences (that I mentioned before), in the next weeks. I'm sorry for taking this long to reply, I had a lot of work for my PhD, a lot of travelling.

Also, I want to thank all of you, @Bisaloo @RomainAzais @Athene-ai for your contribution to this work in the past 6 months. Warm wishes to all of you for the holy season!

Best regards,
Ioannis
