### 2024-01-18 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
Hi @Mavrogiannis-Ioannis, thanks for checking!

I agree it's good enough at this point but I strongly encourage you to revisit in the future. A proper testing framework such as the testthat R package, and in particular [this category of functions](https://testthat.r-lib.org/reference/expect_error.html), would help you write tests more easily.

> So it makes sense that the above are not reached by the examples; the examples are well-constructed, and showcase the normal usage, not explicitly the extreme testing under the hood.

I understand your approach but there is value in testing edge cases as well. You do want to verify that the function returns 1. an error with 2. a **sensible** error message when incorrect input is passed by the user.
It's easy, and no so uncommon, to trip here because of typos, etc. Something that you want to result in an error may actually slip through without you noticing.
