### 2024-02-11 #5572 @Bisaloo
Hi all :wave:, let's try to move this submission forward:

- @strevezas, are you able to provide a timeline for your review?
- @Mavrogiannis-Ioannis, you mentioned in https://github.com/openjournals/joss-reviews/issues/5572#issuecomment-1825680051 you wanted to implement simulation of multiple trajectories. Do you have a timeline for this new feature?

Thanks a lot to everyone for your continue work on this submission. I understand it overlaps with other commitments but in my experience, it's good to try and get it out the door so you can remove it from your mind once and for all and move on :relaxed:.
