### 2024-02-11 #5572 @Mavrogiannis-Ioannis
Hello everyone!
I tried to implement the multiple trajectories, but it turns out that some theoretical work is still required,
so it will be implemented in the future (perhaps during the next year or two).

So, we keep going with what we have right now!
Thank you everyone for your work so far -- I am waiting for further comments and suggestions.

Best regards,
    Ioannis
