### 2024-02-24 #5572 @strevezas
Dear @Bisaloo ,

 I have completed the revision of the paper. According to the package's github page and after running the relevant examples, almost everything seems to be in order.
I have the following remarks:

1. Concerning the automated tests, already mentioned on the issue opened by you, should I check the relevant box? (https://github.com/Mavrogiannis-Ioannis/dsmmR/issues/2)
2. A code of conduct should be added, for example using the usethis::use_tidy_coc() function.

Best regards,
Samis
