### 2024-03-08 Mavrogiannis-Ioannis/dsmmR#2 @Bisaloo
Hi @Mavrogiannis-Ioannis, reviewing this issue after @strevezas' comments, I realize that you do check that functions run without erroring but you do not check that they return the correct output, and that this output is not accidentally modified from one version to the other.

I think this is a serious issue enough that it needs to be fixed before acceptance.
