### 2024-04-07 9172ce4
Create tests for create_sequence()
