### 2024-04-07 b2a3589
Create tests for parametric_dsmm() and non_parametric_dsmm()
