+++ tests/testthat/test-nonparametric_dsmm.R
@@ -1,0 +1,194 @@
+# Setup.
+states <- c("AA", "AC", "CC")
+s <- length(states)
+d <- 2
+k_max <- 3
+
+# `p_dist` has dimensions of: (s, s, d + 1).
+# Sums over v must be 1 for all u and i = 0, ..., d.
+p_dist_1 <- matrix(c(0,   0.1, 0.9,
+                     0.5, 0,   0.5,
+                     0.3, 0.7, 0),
+                   ncol = s, byrow = TRUE)
+
+p_dist_2 <- matrix(c(0,   0.6, 0.4,
+                     0.7, 0,   0.3,
+                     0.6, 0.4, 0),
+                   ncol = s, byrow = TRUE)
+
+p_dist_3 <- matrix(c(0,   0.2, 0.8,
+                     0.6, 0,   0.4,
+                     0.7, 0.3, 0),
+                   ncol = s, byrow = TRUE)
+
+# `f_dist` has dimensions of: (s, s, k_max, d + 1).
+# First f distribution. Dimensions: (s, s, k_max).
+# Sums over l must be 1, for every u, v and i = 0, ..., d.
+f_dist_1_l_1 <- matrix(c(0,   0.2, 0.7,
+                         0.3, 0,   0.4,
+                         0.2, 0.8, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_1_l_2 <- matrix(c(0,   0.3,  0.2,
+                         0.2, 0,    0.5,
+                         0.1, 0.15, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_1_l_3 <- matrix(c(0,   0.5,  0.1,
+                         0.5, 0,    0.1,
+                         0.7, 0.05, 0),
+                       ncol = s, byrow = TRUE)
+# Get f_dist_1
+f_dist_1 <- array(c(f_dist_1_l_1, f_dist_1_l_2, f_dist_1_l_3),
+                  dim = c(s, s, k_max))
+
+# Second f distribution. Dimensions: (s, s, k_max)
+f_dist_2_l_1 <- matrix(c(0,   1/3, 0.4,
+                         0.3, 0,   0.4,
+                         0.2, 0.1, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_2_l_2 <- matrix(c(0,   1/3, 0.4,
+                         0.4, 0,   0.2,
+                         0.3, 0.4, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_2_l_3 <- matrix(c(0,   1/3, 0.2,
+                         0.3, 0,   0.4,
+                         0.5, 0.5, 0),
+                       ncol = s, byrow = TRUE)
+
+# Get f_dist_2
+f_dist_2 <- array(c(f_dist_2_l_1, f_dist_2_l_2, f_dist_2_l_3),
+                  dim = c(s, s, k_max))
+
+# Third f distribution. Dimensions: (s, s, k_max)
+f_dist_3_l_1 <- matrix(c(0,    0.3, 0.3,
+                         0.3,  0,   0.5,
+                         0.05, 0.1, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_3_l_2 <- matrix(c(0,   0.2, 0.6,
+                         0.3, 0,   0.35,
+                         0.9, 0.2, 0),
+                       ncol = s, byrow = TRUE)
+
+f_dist_3_l_3 <- matrix(c(0,    0.5, 0.1,
+                         0.4,  0,   0.15,
+                         0.05, 0.7, 0),
+                       ncol = s, byrow = TRUE)
+
+# Get f_dist_3
[…]
