### 2024-04-07 006d380
Monitor code coverage continuously
