### 2024-04-26 #5572 @Bisaloo
Hi @Mavrogiannis-Ioannis :wave:, this submission is almost through. Could you share a timeline for the last few outstanding points please:

- for the tests, I went ahead and submitted https://github.com/Mavrogiannis-Ioannis/dsmmR/pull/4 to move this forwards. All you need to do now is:
  - [ ] fix the failing test
  - [ ] fix the example the failing test comes from
  - [ ] review and merge the PR
- [ ] for the code of conduct, as mentioned by @strevezas, a simple option is to use the [Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/), which is added in a single line of code by running `usethis::use_code_of_conduct()`.
