+++ R/utils.R
@@ -120,16 +120,23 @@ is_prob_vector <- function(prob) {
 # ==============================================================================
 # Check for the valid use of the different arguments that the user can input
 # ==============================================================================
-valid_sequence <- function(sequence, s) {
+valid_sequence <- function(sequence, s, seq_id = NULL) {
     # '''
     #    Used in `fit_dsmm`.
     # '''
+    warning_message = ''
+    if (!is.null(seq_id)) warning_message <- paste0("In the list `sequences`, ", 
+                                                    "for the sequence numbered (",
+                                                    seq_id,
+                                                    "), this warning occured:")
     if (!is.character(sequence)){
-        stop("\nThe `sequence` argument should be a character vector.")
+        stop(paste0(warning_message, 
+                    "\nThe `sequence` argument should be a character vector."))
     } else if (!missing(s) && length(unique(sequence)) < s) {
         # check for (very) short sequence.
-        stop("\nThe `sequence` argument should have more than s = ",
-             s, " values.")
+        stop(paste0(warning_message, 
+                    "\nThe `sequence` argument should have more than s = ",
+                    s, " values."))
     }
     TRUE
 }
@@ -343,6 +350,9 @@ valid_estimation <- function(estimation, fpar, s, f_is_drifting, degree,
 valid_p_dist <- function(p_dist, s, degree, p_is_drifting, states) {
     # '''
     #    Intended for use in `dsmm_nonparametric` & `dsmm_parametric`.
+    #    The parametric case only concerns the sojourn dime distribution, f.
+    #    In other words, we do not need to separate cases for `valid_p_dist`,
+    #    only the non-parametric case of p is concerned.
     # '''
     D <- degree + 1L
     if (p_is_drifting) {
@@ -367,9 +377,10 @@ valid_p_dist <- function(p_dist, s, degree, p_is_drifting, states) {
             stop("\nThe diagonal values of `p_dist` should all be",
                  " equal to 0.")
         } else if (!all_equal_numeric(
-            (p_drift_rowsum <-
-             c(sapply(1:D, function(u) rowSums(p_dist[ , , u])))),
-            1)) {
+            (p_drift_rowsum <- c(sapply(1:D,
+                                        function(u) rowSums(p_dist[ , , u])))),
+            no_diag <- sapply(p_drift_rowsum, function(i) if (i > 0) 1 else 0))
+        ) {
             # Case where, for every d + 1 matrix, the sums over v for every u
             #  are not equal to 1.
             logical_vector <- sapply(p_drift_rowsum,
@@ -471,10 +482,13 @@ valid_fdist_nonparametric <- function(f_dist, states, s, degree,
             stop("\nThe diagonal values of `f_dist` should all be equal to 0.")
         } else if (!all_equal_numeric(
             f_drift_l_sum <- apply(f_dist, c(1,2,4), sum),
-            no_diag <- array(array(1, dimension[1:2]) -
-                             base::diag(dimension[1]),
-                             dim = dimension[-3]))
-        ) { # Sums over l are not equal to 1.
+            no_diag <- sapply(f_drift_l_sum,
+                              function(i) if (i > 0) 1 else 0)
+                       # array(array(1, dimension[1:2]) -
+                       #       base::diag(dimension[1]),
+                       #       dim = dimension[-3]))
+        )) {
+            # Sums over l are not equal to 1.
             array_names <- sapply(
[…]
