### 2024-06-24 ef7caf4
adding a set.seed(1) argument in tests-fit_dsmm.R
