+++ tests/testthat/_snaps/fit_dsmm.md
@@ -1,1161 +1,0 @@
-# fit_dsmm() non parametric estimation; p and f are drifting
-
-    Code
-      obj_model_1
-    Output
-      
-      
-      
-      $dist$p_drift
-      
-      , , p_0
-      
-                a         c         g         t
-      a 0.0000000 0.3256009 0.3389588 0.3354403
-      c 0.3083129 0.0000000 0.4001558 0.2915313
-      g 0.3630018 0.4284857 0.0000000 0.2085125
-      t 0.1689586 0.2959210 0.5351205 0.0000000
-      
-      , , p_(1/3)
-      
-                a         c         g         t
-      a 0.0000000 0.3180884 0.3353135 0.3465981
-      c 0.3471257 0.0000000 0.4012579 0.2516163
-      g 0.3200541 0.3858379 0.0000000 0.2941080
-      t 0.2220823 0.2890988 0.4888189 0.0000000
-      
-      , , p_(2/3)
-      
-                a         c         g         t
-      a 0.0000000 0.2676607 0.2836909 0.4486484
-      c 0.3991779 0.0000000 0.2596835 0.3411386
-      g 0.3336588 0.3504206 0.0000000 0.3159207
-      t 0.2931447 0.3479670 0.3588883 0.0000000
-      
-      , , p_1
-      
-                a         c         g         t
-      a 0.0000000 0.3107121 0.3430732 0.3462146
-      c 0.3620579 0.0000000 0.4110977 0.2268445
-      g 0.3790501 0.3404757 0.0000000 0.2804741
-      t 0.2627210 0.2521725 0.4851065 0.0000000
-      
-      
-      
-      
-      $dist$f_drift
-      
-      , , l = 1, f_0
-      
-                a         c         g         t
-      a 0.0000000 0.6792766 0.7068605 0.8329514
-      c 0.7479895 0.0000000 0.6218986 0.8122561
-      g 0.7157789 0.7433457 0.0000000 0.6871453
-      t 0.6188977 0.6905871 0.8938019 0.0000000
-      
-      , , l = 2, f_0
-      
-                a         c          g          t
-      a 0.0000000 0.2230097 0.14002987 0.09500086
-      c 0.2140188 0.0000000 0.26273659 0.14439065
-      g 0.2148434 0.2043045 0.00000000 0.22848674
-      t 0.2028256 0.2122140 0.07821943 0.00000000
-      
-      , , l = 3, f_0
-      
-                 a          c          g          t
-      a 0.00000000 0.03216624 0.10730434 0.03802709
-      c 0.02892532 0.00000000 0.10551129 0.02527413
-      g 0.05591723 0.04742578 0.00000000 0.07427857
-      t 0.14715493 0.05679444 0.01007655 0.00000000
-      
-      , , l = 4, f_0
-      
-                  a          c           g           t
-      a 0.000000000 0.04793010 0.032508871 0.026585731
-      c 0.008833648 0.00000000 0.008976874 0.012098460
-      g 0.009766701 0.00299179 0.000000000 0.008101114
-      t 0.014426001 0.03580043 0.003461638 0.000000000
-      
-      , , l = 5, f_0
-      
-                   a           c            g           t
-      a 0.0000000000 0.005201014 0.0116438484 0.002836745
-      c 0.0002327449 0.000000000 0.0008712519 0.005980701
-      g 0.0032516801 0.001713705 0.0000000000 0.001034821
-      t 0.0075716663 0.001225007 0.0133769812 0.000000000
-      
-      , , l = 6, f_0
-      
[…]

[…]
