### 2024-06-24 85c4fa4
adding contributor covenant code of conduct badge
