### 2024-06-24 fb062a1
adding set.seed(1) at parametric estimation functions for equivalent snapshots across macos, linux, etc.
