+++ tests/testthat/_snaps/fit_dsmm.md
@@ -1036,7 +1036,7 @@
       
       Class: dsmm_fit_parametric, dsmm
 
----
+# fit_dsmm() parametric estimation; p is drifting, f is not drifting
 
     Code
       obj_fit_parametric_2
@@ -1152,6 +1152,177 @@
       
       $Model
       Model_2 
+      
+      $estimation
+      parametric 
+      
+      
+      Class: dsmm_fit_parametric, dsmm
+
+# fit_dsmm() parametric estimation; p is not drifting, f is drifting
+
+    Code
+      obj_fit_parametric_3
+    Output
+      
+      
+      
+      $dist$p_notdrift
+      
+                a         c         g         t
+      a 0.0000000 0.3055155 0.3252591 0.3692254
+      c 0.3541686 0.0000000 0.3680487 0.2777827
+      g 0.3489412 0.3763050 0.0000000 0.2747538
+      t 0.2367266 0.2962898 0.4669836 0.0000000
+      
+      
+      
+      $dist$f_drift_parametric
+      
+      , , 1
+      
+           [,1]       [,2]   [,3]       [,4]      
+      [1,] NA         "unif" "dweibull" "nbinom"  
+      [2,] "pois"     NA     "pois"     "dweibull"
+      [3,] "geom"     "pois" NA         "geom"    
+      [4,] "dweibull" "geom" "pois"     NA        
+      
+      , , 2
+      
+           [,1]       [,2]   [,3]       [,4]      
+      [1,] NA         "unif" "dweibull" "nbinom"  
+      [2,] "pois"     NA     "pois"     "dweibull"
+      [3,] "geom"     "pois" NA         "geom"    
+      [4,] "dweibull" "geom" "pois"     NA        
+      
+      , , 3
+      
+           [,1]       [,2]   [,3]       [,4]      
+      [1,] NA         "unif" "dweibull" "nbinom"  
+      [2,] "pois"     NA     "pois"     "dweibull"
+      [3,] "geom"     "pois" NA         "geom"    
+      [4,] "dweibull" "geom" "pois"     NA        
+      
+      , , 4
+      
+           [,1]       [,2]   [,3]       [,4]      
+      [1,] NA         "unif" "dweibull" "nbinom"  
+      [2,] "pois"     NA     "pois"     "dweibull"
+      [3,] "geom"     "pois" NA         "geom"    
+      [4,] "dweibull" "geom" "pois"     NA        
+      
+      
+      
+      
+      $dist$f_drift_parameters
+      
+      , , 1, fpars_0
+      
+                a         c         g         t
+      a        NA 4.0000000 0.2931395 0.2573780
+      c 0.2993014        NA 0.5042017 0.1877439
+      g 0.7292906 0.3162102        NA 0.7090922
+      t 0.3811023 0.6864256 0.1678208        NA
+      
+      , , 2, fpars_0
+      
+                a  c         g         t
+      a        NA NA 0.8820125 0.4702177
+      c        NA NA        NA 1.3183296
+      g        NA NA        NA        NA
+      t 0.9853349 NA        NA        NA
+      
+      , , 1, fpars_(1/3)
+      
+                a         c         g         t
+      a        NA 5.0000000 0.2323796 0.4508238
+      c 0.2130453        NA 0.3619084 0.2117913
+      g 0.7262501 0.4018156        NA 0.7563290
+      t 0.2716859 0.6912428 0.2372413        NA
+      
+      , , 2, fpars_(1/3)
+      
[…]
