### 2024-06-24 b128b1f
switching to a random sequence instead of the data Lambda for the test case of `fit_dsmm`
