+++ tests/testthat/_snaps/fit_dsmm.md
@@ -11,34 +11,34 @@
       , , p_0
       
                 a         c         g         t
-      a 0.0000000 0.3256009 0.3389588 0.3354403
-      c 0.3083129 0.0000000 0.4001558 0.2915313
-      g 0.3630018 0.4284857 0.0000000 0.2085125
-      t 0.1689586 0.2959210 0.5351205 0.0000000
+      a 0.0000000 0.2928374 0.3717104 0.3354522
+      c 0.3023716 0.0000000 0.3962073 0.3014211
+      g 0.3436550 0.3047510 0.0000000 0.3515940
+      t 0.3325151 0.2558372 0.4116477 0.0000000
       
       , , p_(1/3)
       
                 a         c         g         t
-      a 0.0000000 0.3180884 0.3353135 0.3465981
-      c 0.3471257 0.0000000 0.4012579 0.2516163
-      g 0.3200541 0.3858379 0.0000000 0.2941080
-      t 0.2220823 0.2890988 0.4888189 0.0000000
+      a 0.0000000 0.3684005 0.3296269 0.3019726
+      c 0.3520526 0.0000000 0.3391474 0.3088001
+      g 0.3714217 0.3247288 0.0000000 0.3038496
+      t 0.3554353 0.3402464 0.3043183 0.0000000
       
       , , p_(2/3)
       
                 a         c         g         t
-      a 0.0000000 0.2676607 0.2836909 0.4486484
-      c 0.3991779 0.0000000 0.2596835 0.3411386
-      g 0.3336588 0.3504206 0.0000000 0.3159207
-      t 0.2931447 0.3479670 0.3588883 0.0000000
+      a 0.0000000 0.3088205 0.3285419 0.3626376
+      c 0.3434957 0.0000000 0.3389797 0.3175246
+      g 0.3208303 0.3117509 0.0000000 0.3674188
+      t 0.3407074 0.2749527 0.3843399 0.0000000
       
       , , p_1
       
                 a         c         g         t
-      a 0.0000000 0.3107121 0.3430732 0.3462146
-      c 0.3620579 0.0000000 0.4110977 0.2268445
-      g 0.3790501 0.3404757 0.0000000 0.2804741
-      t 0.2627210 0.2521725 0.4851065 0.0000000
+      a 0.0000000 0.3918856 0.3910772 0.2170372
+      c 0.3454498 0.0000000 0.3041657 0.3503845
+      g 0.3226336 0.3867389 0.0000000 0.2906275
+      t 0.3444807 0.3937185 0.2618008 0.0000000
       
       
       
@@ -48,278 +48,246 @@
       , , l = 1, f_0
       
                 a         c         g         t
-      a 0.0000000 0.6792766 0.7068605 0.8329514
-      c 0.7479895 0.0000000 0.6218986 0.8122561
-      g 0.7157789 0.7433457 0.0000000 0.6871453
-      t 0.6188977 0.6905871 0.8938019 0.0000000
+      a 0.0000000 0.7901328 0.6937341 0.9168894
+      c 0.6063477 0.0000000 0.8197100 0.7144069
+      g 0.6644908 0.7497707 0.0000000 0.7786443
+      t 0.8231650 0.7519020 0.7267143 0.0000000
       
       , , l = 2, f_0
       
-                a         c          g          t
-      a 0.0000000 0.2230097 0.14002987 0.09500086
-      c 0.2140188 0.0000000 0.26273659 0.14439065
-      g 0.2148434 0.2043045 0.00000000 0.22848674
-      t 0.2028256 0.2122140 0.07821943 0.00000000
+                a         c         g         t
+      a 0.0000000 0.1795942 0.2045875 0.0590463
[…]
