+++ tests/testthat/_snaps/fit_dsmm.md
@@ -815,35 +815,35 @@
       
       , , 1
       
-           [,1]       [,2]   [,3]       [,4]      
-      [1,] NA         "unif" "dweibull" "nbinom"  
-      [2,] "pois"     NA     "pois"     "dweibull"
-      [3,] "geom"     "pois" NA         "geom"    
-      [4,] "dweibull" "geom" "pois"     NA        
+           [,1]   [,2]   [,3]   [,4]  
+      [1,] NA     "unif" "geom" "pois"
+      [2,] "pois" NA     "pois" "geom"
+      [3,] "geom" "pois" NA     "geom"
+      [4,] "geom" "geom" "pois" NA    
       
       , , 2
       
-           [,1]       [,2]   [,3]       [,4]      
-      [1,] NA         "unif" "dweibull" "nbinom"  
-      [2,] "pois"     NA     "pois"     "dweibull"
-      [3,] "geom"     "pois" NA         "geom"    
-      [4,] "dweibull" "geom" "pois"     NA        
+           [,1]   [,2]   [,3]   [,4]  
+      [1,] NA     "unif" "geom" "pois"
+      [2,] "pois" NA     "pois" "geom"
+      [3,] "geom" "pois" NA     "geom"
+      [4,] "geom" "geom" "pois" NA    
       
       , , 3
       
-           [,1]       [,2]   [,3]       [,4]      
-      [1,] NA         "unif" "dweibull" "nbinom"  
-      [2,] "pois"     NA     "pois"     "dweibull"
-      [3,] "geom"     "pois" NA         "geom"    
-      [4,] "dweibull" "geom" "pois"     NA        
+           [,1]   [,2]   [,3]   [,4]  
+      [1,] NA     "unif" "geom" "pois"
+      [2,] "pois" NA     "pois" "geom"
+      [3,] "geom" "pois" NA     "geom"
+      [4,] "geom" "geom" "pois" NA    
       
       , , 4
       
-           [,1]       [,2]   [,3]       [,4]      
-      [1,] NA         "unif" "dweibull" "nbinom"  
-      [2,] "pois"     NA     "pois"     "dweibull"
-      [3,] "geom"     "pois" NA         "geom"    
-      [4,] "dweibull" "geom" "pois"     NA        
+           [,1]   [,2]   [,3]   [,4]  
+      [1,] NA     "unif" "geom" "pois"
+      [2,] "pois" NA     "pois" "geom"
+      [3,] "geom" "pois" NA     "geom"
+      [4,] "geom" "geom" "pois" NA    
       
       
       
@@ -852,67 +852,67 @@
       
       , , 1, fpars_0
       
-                a         c         g          t
-      a        NA 3.0000000 0.3062659 0.08844116
-      c 0.5526086        NA 0.2642541 0.28559307
-      g 0.6821648 0.2882907        NA 0.76730322
-      t 0.1768350 0.7984301 0.3492194         NA
+                a         c         g         t
+      a        NA 3.0000000 0.6970727 0.1541034
+      c 0.5526086        NA 0.2642541 0.6688289
+      g 0.6821648 0.2882907        NA 0.7673032
+      t 0.8120297 0.7984301 0.3492194        NA
       
       , , 2, fpars_0
       
-                a  c        g         t
-      a        NA NA 1.479848 0.3646389
-      c        NA NA       NA 0.7718552
-      g        NA NA       NA        NA
-      t 0.9199561 NA       NA        NA
+         a  c  g  t
+      a NA NA NA NA
+      c NA NA NA NA
+      g NA NA NA NA
+      t NA NA NA NA
       
       , , 1, fpars_(1/3)
       
                 a         c         g         t
-      a        NA 5.0000000 0.2481787 1.0251200
[…]
