+++ .github/workflows/codecov.yml
@@ -1,4 +1,5 @@
-- name: Upload coverage reports to Codecov
+name: Upload coverage reports to Codecov
+on: [push, pull_request]
   uses: codecov/codecov-action@v4.0.1
   with:
     token: ${{ secrets.CODECOV_TOKEN }}
