+++ .github/workflows/test-coverage.yaml
@@ -42,12 +42,12 @@ jobs:
           find '${{ runner.temp }}/package' -name 'testthat.Rout*' -exec cat '{}' \; || true
         shell: bash
 
-      - name: Upload test results
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: coverage-test-failures
-          path: ${{ runner.temp }}/package
+      # - name: Upload test results
+      #   if: failure()
+      #   uses: actions/upload-artifact@v4
+      #   with:
+      #     name: coverage-test-failures
+      #     path: ${{ runner.temp }}/package
           
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4.0.1
