### 2024-06-24 b2670b3
small update for .gitignore, .Rbuildignore, NEWS.md
