### 2024-06-24 #5572 @Mavrogiannis-Ioannis
Good morning everyone,

Some progress has been made.
Notably,
- Code of Conduct now has been added according to your suggestions.
- Tests have been implemented for the main functions of the package
  (I will merge the issues and pull requests in the next week @Bisaloo )
- codecov github workflow has been established
- Error fixed when printing a `fit_dsmm()` object, which specifically happened only
   in the parametric estimation case, while considering f not drifting (Model 2 case).

Therefore it seems to me that the main questions addressed above are dealt with.
Do not hesitate to offer any other questions, ideas or suggestions.

Best regards,
Ioannis
