### 2024-06-24 d9a9be7
added errors fixed in NEWS
