### 2024-07-11 #5572 @Bisaloo
Hi :wave: @Mavrogiannis-Ioannis, thanks for the changes, and in particular the code of conduct and the tests. I also confirmed the `print()` issue I identified has been fixed.

@Athene-ai @RomainAzais @strevezas, could you please have a last look and confirm all seems good to go from your end? As a reminder, here are the links to your respective review comments, where the main issues seemed to be around the presence of tests:

- @Athene-ai: https://github.com/openjournals/joss-reviews/issues/5572#issuecomment-1597297271
- @RomainAzais: https://github.com/openjournals/joss-reviews/issues/5572#issuecomment-1620244875
- @strevezas: https://github.com/openjournals/joss-reviews/issues/5572#issuecomment-1961801665
