### 2024-07-20 902f652
Add missing DOIs
