+++ paper.bib
@@ -4,6 +4,7 @@
   volume={191},
   year={2009},
   publisher={Springer Science \& Business Media}
+  doi={10.1007/978-0-387-73173-5}
 }
 
 
@@ -20,6 +21,7 @@
     year = {2021},
     note = {R package version 1.0.3},
     url = {https://CRAN.R-project.org/package=smmR},
+    doi = {10.32614/CRAN.package.smmR}
 }
 
 
@@ -54,6 +56,7 @@
     year = {2021},
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=drimmR},
+    doi = {10.32614/CRAN.package.drimmR}
 }
 
 
@@ -75,16 +78,17 @@
     number = {6},
     pages = {1--16},
     url = {http://www.jstatsoft.org/v66/i06/},
+    doi = {10.18637/jss.v066.i06}
   }
 
 
 @Manual{HMM_package,
   title = {HMM: Hidden Markov Models},
-  journal = {Scientific Software},
   author = {Lin Himmelmann},
   year = {2022},
   note = {R package version 1.0.1},
   url = {https://CRAN.R-project.org/package=HMM},
+  doi = {10.32614/CRAN.package.HMM}
 }
 
 @Manual{HiddenMarkov_package,
@@ -95,6 +99,7 @@
     organization = {Statistics Research Associates},
     address = {Wellington},
     url = {https://www.statsresearch.co.nz/dsh/sslib/},
+    doi = {10.32614/CRAN.package.HiddenMarkov}
   }
 
 @Article{mhsmm_package,
@@ -106,4 +111,5 @@
     number = {4},
     pages = {1--22},
     url = {http://www.jstatsoft.org/v39/i04/},
+    doi = {10.18637/jss.v039.i04}
   }
