### 2024-07-20 f4a2692
Fix citation format
