+++ paper.md
@@ -29,7 +29,7 @@ editor_options:
 
 Markov models are a common approach in the researcher's toolbox in order to deal with the modeling of many real-life phenomenons presented as a sequence under discrete time. However, this imposes the assumption that the sequence under inspection is homogeneous with respect to time and furthermore that the sojourn times follow the Geometric distribution. This is not always true in practice when modeling, for example, DNA sequences. 
 
-Drifting semi-Markov models (DSMM) formulate the combination of semi-Markov models, which allow arbitrary distributions for the sojourn times, together with drifting Markov models, which describe the non-homogeneity of a sequence through a smooth, known shape that is gradually evolving, expressed through a linear (polynomial) function. As a result, DSMM are best suited to capture the non-homogeneity that occurs from the intrinsic evolution of the system or from the interactions between the system and the environment. For a detailed introduction to semi-Markov models see [@barbu_limnios]. Drifting Markov models were first introduced in [@drift_polynomial].
+Drifting semi-Markov models (DSMM) formulate the combination of semi-Markov models, which allow arbitrary distributions for the sojourn times, together with drifting Markov models, which describe the non-homogeneity of a sequence through a smooth, known shape that is gradually evolving, expressed through a linear (polynomial) function. As a result, DSMM are best suited to capture the non-homogeneity that occurs from the intrinsic evolution of the system or from the interactions between the system and the environment. For a detailed introduction to semi-Markov models see @barbu_limnios. Drifting Markov models were first introduced in @drift_polynomial.
 
 `dsmmR` is an R package which allows the user to perform parametric and non-parametric estimation and simulation of drifting semi-Markov processes. The user can also define their own parametric and non-parametric DSMM specifications, allowing for a necessary degree of freedom when dealing with a research question. Furthermore, three different types of DSMM are considered. These three models differ in the way they characterize the drifting semi-Markov kernel. Specifically, the first model allows both the transition matrix and the sojourn time distribution to vary (i.e. to "drift"), the second model allows only the transition matrix to drift while the third model allows only the sojourn time distribution to drift.
 
