### 2024-07-20 #5572 @Bisaloo
@Mavrogiannis-Ioannis, I have confirmed myself that the remaining uncheck boxes are now complete. Since they are very practical & objective points (presence of a code of conduct, presence of tests, etc.), I'm confident with moving forward with the acceptance of this submission.

I have proofread the article and submitted a pull request to fix minor formatting issues, and add the missing DOIs: https://github.com/Mavrogiannis-Ioannis/dsmmR/pull/5.

Once the PR is merged, could you please:

- [ ] Make a last proofreading pass from your end since the article will be published directly after acceptance by the editor in chief
