+++ paper.bib
@@ -3,7 +3,7 @@
   author={Barbu, Vlad Stefan and Limnios, Nikolaos},
   volume={191},
   year={2009},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
   doi={10.1007/978-0-387-73173-5}
 }
 
