### 2024-07-20 5417af8
Updated clarity for README + warning messages. Also updated the paper for JOSS, trying to put also some code in a Quickstart section.
