+++ paper.md
@@ -11,14 +11,18 @@ authors:
     orcid: 0000-0002-0840-016X
     affiliation: 1 
   - name: Ioannis Mavrogiannis
-    affiliation: 1
+    affiliation: 2
     corresponding: true
   - name: Nicolas Vergne
     affiliation: 1
 affiliations:
-  - name: Univ Rouen Normandie, CNRS, LMRS UMR 6085, F-76000 Rouen, France
+  - name: Laboratory of Mathematics Raphaël Salem (LMRS), UMR CNRS 6085, University of Rouen Normandy, France
     index: 1
-date: 22 January 2023
+  - name: Laboratoire de Mécanique, Modélisation & Procédés Propres (M2P2), M2P2 UMR 7340, Aix Marseille Université, CNRS, Centrale Méditerranée, France 
+    index: 2
+  - name: Univ Rouen Normandie, CNRS, LMRS UMR 6085, F-76000 Rouen, France
+    index: 3
+date: 20 July 2024
 bibliography: paper.bib
 editor_options: 
   markdown: 
@@ -27,7 +31,7 @@ editor_options:
 
 # Summary
 
-Markov models are a common approach in the researcher's toolbox in order to deal with the modeling of many real-life phenomenons presented as a sequence under discrete time. However, this imposes the assumption that the sequence under inspection is homogeneous with respect to time and furthermore that the sojourn times follow the Geometric distribution. This is not always true in practice when modeling, for example, DNA sequences. 
+Markov models are a common approach in the researcher's toolbox in order to deal with the modeling of many real-life phenomenons presented as a sequence under discrete time. However, this imposes the assumption that the sequence under inspection is homogeneous with respect to time and furthermore that the sojourn times follow the Geometric distribution. This is not always true in practice when modeling, for example, DNA sequences.
 
 Drifting semi-Markov models (DSMM) formulate the combination of semi-Markov models, which allow arbitrary distributions for the sojourn times, together with drifting Markov models, which describe the non-homogeneity of a sequence through a smooth, known shape that is gradually evolving, expressed through a linear (polynomial) function. As a result, DSMM are best suited to capture the non-homogeneity that occurs from the intrinsic evolution of the system or from the interactions between the system and the environment. For a detailed introduction to semi-Markov models see @barbu_limnios. Drifting Markov models were first introduced in @drift_polynomial.
 
@@ -56,6 +60,170 @@ Drifting semi-Markov processes represent a versatile tool that can be applied in
[…]
