+++ R/utils.R
@@ -181,7 +181,7 @@ valid_k_max <- function(k_max, soj_times) {
             return(TRUE)
         } else {
             stop("\nThe maximum sojourn time attribute `k_max` should be",
-                 " an integer equal to the maximum of `soj_times`.")
+                 " an integer,\nequal to the maximum of `soj_times`.")
         }
     }
     stop("\nThe maximum sojourn time `k_max` should be an integer.")
@@ -219,8 +219,8 @@ valid_states <- function(states) {
     } else if(min(prob <- (sapply(strsplit(
         x = gsub(" ", "", states), split = ""), length))) == 0) {
         # Check for a state equal to an empty string "" or `character(0)`.
-        stop("\nState space `states` includes a state without a given name, ",
-             "at position ", which(prob == 0))
+        stop("\nState space `states` includes a state without a given name,\n",
+             "at position(s): ", which(prob == 0))
     }
     TRUE
 }
@@ -234,8 +234,8 @@ valid_state <- function(state, states) {
     }
     stopifnot(valid_states(states))
     if (!state %in% states) {
-        stop("\nState `", state, "` is not included in the State",
-             " Space that is given,\nE = (", paste(states, collapse = ', '),
+        stop("\nState `", state, "` is not included in the state",
+             " space that is given,\nE = (", paste(states, collapse = ', '),
              ").")
     }
     TRUE
@@ -247,7 +247,7 @@ valid_length_states <- function(s, states) {
     # '''
     if (!(is_integer(s) && s == (lstates <- length(states)))) {
         stop("\nAttribute `s` = ", s,
-             " should be an integer equal to the length of `states` = ",
+             " should be an integer,\nequal to the length of `states` = ",
              lstates)
     }
     TRUE
@@ -255,7 +255,7 @@ valid_length_states <- function(s, states) {
 
 valid_degree <- function(degree, model_size) {
     # '''
-    #   This functions checks for correct behavior of the `degree` parameter.
+    #   This functions checks for correct behaviour of the `degree` parameter.
     # '''
     if (!is_integer(degree)) {
         stop("\nThe the polynomial `degree` should be a positive integer.")
@@ -263,8 +263,9 @@ valid_degree <- function(degree, model_size) {
     if (!missing(model_size)) {
         if (degree > model_size) {
             stop("\nThe polynomial `degree` = ", degree,
-                 " should not be larger than the model size = ",
-                 model_size, ".")
+                 " should not be larger than the model size = ", model_size,
+                 ",\nwhich is the length of the EMC of the sequence, minus 1.",
+                 "\nSee `Definition` section at ?dsmmR for more information.")
         }
     }
     TRUE
@@ -313,9 +314,9 @@ valid_model <- function(p_is_drifting, f_is_drifting) {
     # '''
     if (!p_is_drifting && !f_is_drifting) {
         # Neither p or f are drifting.
-        stop("\nAt least p or f should be drifting to use a drifting ",
-             "semi-Markov model for the estimation. Otherwise, ",
-             "a semi-Markov model should be used.")
+        stop("\nAt least p or f should be drifting,\nto use a drifting ",
+             "semi-Markov model for the estimation.\n", 
+             "Otherwise, a semi-Markov model should be used.")
     }
     TRUE
 }
@@ -330,13 +331,13 @@ valid_estimation <- function(estimation, fpar, s, f_is_drifting, degree,
     # '''
     if (!is.character(estimation) ||
         (estimation != "nonparametric" && estimation != "parametric")) {
[…]
