+++ paper.md
@@ -63,67 +63,12 @@ Furthermore, for drifting Markov models the R package `drimmR` [@drimmR_package]
 
 # Quickstart
 
-```{r, eval = TRUE, include = FALSE}
-library(dsmmR)
-states <- c("a", "b", "c")
-s <- length(states)
-degree <- 1
-p_dist_1 <- matrix(c(0,   0.4,  0.6,
-                     0.5, 0,    0.5,
-                     0.3, 0.7,  0   ), ncol = s, byrow = TRUE)
-p_dist_2 <- matrix(c(0,   0.55, 0.45,
-                     0.25, 0,   0.75,
-                     0.5, 0.5,  0   ), ncol = s, byrow = TRUE)
-p_dist <- array(c(p_dist_1, p_dist_2), dim = c(s, s, degree + 1))
-f_dist_1 <- matrix(c(NA,   "nbinom",   "unif",
-                   "geom",  NA,        "pois",
-                   "pois", "dweibull",  NA   ), nrow = s, ncol = s, byrow = TRUE)
-f_dist_1_pars_1 <- matrix(c(NA,  4,   3,
-                            0.7, NA,  5,
-                            3,   0.6, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_1_pars_2 <- matrix(c(NA,  0.5, NA,
-                            NA,  NA,  NA,
-                            NA,  0.8, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_2 <- f_dist_1 
-f_dist_2_pars_1 <- matrix(c(NA,  3,   5,
-                            0.3, NA,  2,
-                            5,   0.3, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_2_pars_2 <- matrix(c(NA,  0.4, NA,
-                            NA,  NA,  NA,
-                            NA,  0.5, NA), nrow = s, ncol = s, byrow = TRUE)`.
-
-f_dist <- array(c(f_dist_1, f_dist_2), dim = c(s, s, degree + 1))
-f_dist_pars <- array(c(f_dist_1_pars_1, f_dist_1_pars_2,
-                       f_dist_2_pars_1, f_dist_2_pars_2), 
-                     dim = c(s, s, 2, degree + 1))
-dsmm_model <- parametric_dsmm(
-    model_size = 10000,
-    states = states,
-    initial_dist = c(0.6, 0.3, 0.1),
-    degree = degree,
-    p_dist = p_dist,
-    f_dist = f_dist,
-    f_dist_pars = f_dist_pars,
-    p_is_drifting = TRUE,
-    f_is_drifting = TRUE
-)
-sim_seq <- simulate(dsmm_model, klim = 30, seed = 1)
-fitted_model <- fit_dsmm(sequence = sim_seq,
-                         states = states,
-                         degree = degree,
-                         f_is_drifting = TRUE,
-                         p_is_drifting = TRUE,
-                         estimation = 'parametric',
-                         f_dist = f_dist)
-```
-
-
 To install the package, simply use either of the commands in the R console:
-```{r, eval = FALSE}
+```r
 install.packages('dsmmR') # official release in CRAN
 ```
 Otherwise, if you wish to download the development version directly from Github: 
-```{r, eval = FALSE}
+```r
 if (!require("devtools")) install.packages("devtools")
 devtools::install_github("Mavrogiannis-Ioannis/dsmmR")
 ```
@@ -131,18 +76,18 @@ devtools::install_github("Mavrogiannis-Ioannis/dsmmR")
 We can then set up a simple case of defining a parametric `dsmm` object,
 simulating from it and then estimating from the simulated sequence, 
 by first of all loading the package, 
-```{r, eval = FALSE}
+```r
[…]
