### 2024-07-20 df32b56
remove sections added in paper and add example in README about parametric models
