+++ paper.md
@@ -61,162 +61,6 @@ For semi-Markov models we have the R package `SemiMarkov` [@SemiMarkov_package] 
 
 Furthermore, for drifting Markov models the R package `drimmR` [@drimmR_package] was developed, performing non-parametric estimation and simulation, also allowing for the exact computation of the associated reliability measures and for several frameworks with regards to how many samples were used, considering if they were complete or incomplete and if they were of the same length. In this package, a linear (polynomial) function was used to capture the non-homogeneity. 
 
-# Quickstart
-
-To install the package, simply use either of the commands in the R console:
-```r
-install.packages('dsmmR') # official release in CRAN
-```
-Otherwise, if you wish to download the development version directly from Github: 
-```r
-if (!require("devtools")) install.packages("devtools")
-devtools::install_github("Mavrogiannis-Ioannis/dsmmR")
-```
-
-We can then set up a simple case of defining a parametric `dsmm` object,
-simulating from it and then estimating from the simulated sequence, 
-by first of all loading the package, 
-```r
-library(dsmmR)
-```
-and then defining the states and a degree equal to 1.
-```r
-states <- c("a", "b", "c")
-s <- length(states)
-degree <- 1
-```
-
-Since degree is equal to 2, we can then define the 2 drifting transition matrices:
-```r
-p_dist_1 <- matrix(c(0,   0.4,  0.6,
-                     0.5, 0,    0.5,
-                     0.3, 0.7,  0   ), ncol = s, byrow = TRUE)
-p_dist_2 <- matrix(c(0,   0.55, 0.45,
-                     0.25, 0,   0.75,
-                     0.5, 0.5,  0   ), ncol = s, byrow = TRUE)
-p_dist <- array(c(p_dist_1, p_dist_2), dim = c(s, s, degree + 1))
-```
-
-Let us also consider the case where only the parameters of the distributions 
-modeling the sojourn times are drifting across the sequence. Note that some 
-distributions like the Negative Binomial and the Discrete Weibull require two 
-parameters, which we define in two matrices.
-```r
-f_dist_1 <- matrix(c(NA,   "nbinom",   "unif",
-                   "geom",  NA,        "pois",
-                   "pois", "dweibull",  NA   ), nrow = s, ncol = s, byrow = TRUE)
-f_dist_1_pars_1 <- matrix(c(NA,  4,   3,
-                            0.7, NA,  5,
-                            3,   0.6, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_1_pars_2 <- matrix(c(NA,  0.5, NA,
-                            NA,  NA,  NA,
-                            NA,  0.8, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_2 <- f_dist_1 
-f_dist_2_pars_1 <- matrix(c(NA,  3,   5,
-                            0.3, NA,  2,
-                            5,   0.3, NA), nrow = s, ncol = s, byrow = TRUE)
-f_dist_2_pars_2 <- matrix(c(NA,  0.4, NA,
-                            NA,  NA,  NA,
-                            NA,  0.5, NA), nrow = s, ncol = s, byrow = TRUE).
-
-f_dist <- array(c(f_dist_1, f_dist_2), dim = c(s, s, degree + 1))
-f_dist_pars <- array(c(f_dist_1_pars_1, f_dist_1_pars_2,
-                       f_dist_2_pars_1, f_dist_2_pars_2), 
[…]
