+++ README.Rmd
@@ -111,16 +111,6 @@ fitted_model <- fit_dsmm(sequence = sequence,
 For more details about the estimation, consider viewing the extended
 documentation through `?fit_dsmm`.
 
-### Defining drifting semi-Markov models
-
-When defining a DSMM object we need to input parameters like the
-polynomial degree, the state space, the DSMM size (length of the
-embedded Markov chain), the sojourn times *f*, the transition matrices
-*p* and more.
-
-For more information, consider the documentation through
-`?parametric_dsmm` and `?nonparametric_dsmm`.
-
 ### Simulation
 
 After fitting a DSMM (or defining it through `nonparametric_dsmm()` or
@@ -151,6 +141,160 @@ The dimensionality of the DSM kernel can be reduced further through the
 attributes of the function.
 
 For more information, consider the documentation through `?get_kernel`.
+
+
+### Defining drifting semi-Markov models
+
+When defining a DSMM object we need to input parameters like the
+polynomial degree, the state space, the DSMM size (length of the
+embedded Markov chain), the sojourn times *f*, the transition matrices
+*p* and more.
+
+See example below. 
+
+For more information, consider the documentation through
+`?parametric_dsmm` and `?nonparametric_dsmm`.
+
+We can set up a simple case of defining a parametric `dsmm` object,
+simulating from it and then estimating from the simulated sequence, 
+by first of all loading the package, 
+```r
+library(dsmmR)
+```
+and then defining the states and a degree equal to 1.
+```r
+states <- c("a", "b", "c")
+s <- length(states)
+degree <- 1
+```
+
+Since degree is equal to 1, we can then define the 2 drifting transition matrices:
+```r
+p_dist_1 <- matrix(c(0,   0.4,  0.6,
+                     0.5, 0,    0.5,
+                     0.3, 0.7,  0   ), ncol = s, byrow = TRUE)
+p_dist_2 <- matrix(c(0,   0.55, 0.45,
+                     0.25, 0,   0.75,
+                     0.5, 0.5,  0   ), ncol = s, byrow = TRUE)
+p_dist <- array(c(p_dist_1, p_dist_2), dim = c(s, s, degree + 1))
+```
+
+Let us also consider the case where only the parameters of the distributions 
+modeling the sojourn times are drifting across the sequence. Note that some 
+distributions like the Negative Binomial and the Discrete Weibull require two 
+parameters, which we define in two matrices.
+```r
+f_dist_1 <- matrix(c(NA,   "nbinom",   "unif",
+                   "geom",  NA,        "pois",
+                   "pois", "dweibull",  NA   ), nrow = s, ncol = s, byrow = TRUE)
+f_dist_1_pars_1 <- matrix(c(NA,  4,   3,
+                            0.7, NA,  5,
+                            3,   0.6, NA), nrow = s, ncol = s, byrow = TRUE)
+f_dist_1_pars_2 <- matrix(c(NA,  0.5, NA,
+                            NA,  NA,  NA,
+                            NA,  0.8, NA), nrow = s, ncol = s, byrow = TRUE)
+f_dist_2 <- f_dist_1 
+f_dist_2_pars_1 <- matrix(c(NA,  3,   5,
+                            0.3, NA,  2,
+                            5,   0.3, NA), nrow = s, ncol = s, byrow = TRUE)
+f_dist_2_pars_2 <- matrix(c(NA,  0.4, NA,
[…]

[…]
