### 2024-07-20 06b1124
cleaning README code for new example section
