### 2024-07-20 c2362ce
cleaning README code for new example section
