+++ README.Rmd
@@ -145,31 +145,29 @@ For more information, consider the documentation through `?get_kernel`.
 
 ### Defining drifting semi-Markov models
 
-When defining a DSMM object we need to input parameters like the
-polynomial degree, the state space, the DSMM size (length of the
-embedded Markov chain), the sojourn times *f*, the transition matrices
-*p* and more.
-
-See example below. 
+We can put together all the previous concepts in the showcase of parametric
+estimation. First, we will define the drifting transition matrices and the 
+drifting sojourn time distributions. Then, we will create a `dsmm_parametric` 
+object, we will simulate a sequence from it and then finally we will estimate 
+a drifting semi-Markov model from that simulated sequence.
 
 For more information, consider the documentation through
 `?parametric_dsmm` and `?nonparametric_dsmm`.
 
-We can set up a simple case of defining a parametric `dsmm` object,
-simulating from it and then estimating from the simulated sequence, 
-by first of all loading the package, 
-```r
+First of all we load the package, 
+```{r, eval = FALSE}
 library(dsmmR)
 ```
-and then defining the states and a degree equal to 1.
-```r
+
+and then we define the states and we set the degree equal to 1.
+```{r, eval = FALSE}
 states <- c("a", "b", "c")
 s <- length(states)
 degree <- 1
 ```
 
-Since degree is equal to 1, we can then define the 2 drifting transition matrices:
-```r
+Since degree is equal to 1, we then define the 2 drifting transition matrices:
+```{r, eval = FALSE}
 p_dist_1 <- matrix(c(0,   0.4,  0.6,
                      0.5, 0,    0.5,
                      0.3, 0.7,  0   ), ncol = s, byrow = TRUE)
@@ -180,10 +178,10 @@ p_dist <- array(c(p_dist_1, p_dist_2), dim = c(s, s, degree + 1))
 ```
 
 Let us also consider the case where only the parameters of the distributions 
-modeling the sojourn times are drifting across the sequence. Note that some 
+modeling the sojourn times are drifting across the sequence. Note that 
 distributions like the Negative Binomial and the Discrete Weibull require two 
-parameters, which we define in two matrices.
-```r
+parameters, which we define in two matrices for each distribution.
+```{r, eval = FALSE}
 f_dist_1 <- matrix(c(NA,   "nbinom",   "unif",
                    "geom",  NA,        "pois",
                    "pois", "dweibull",  NA   ), nrow = s, ncol = s, byrow = TRUE)
@@ -199,7 +197,7 @@ f_dist_2_pars_1 <- matrix(c(NA,  3,   5,
                             5,   0.3, NA), nrow = s, ncol = s, byrow = TRUE)
 f_dist_2_pars_2 <- matrix(c(NA,  0.4, NA,
                             NA,  NA,  NA,
-                            NA,  0.5, NA), nrow = s, ncol = s, byrow = TRUE).
+                            NA,  0.5, NA), nrow = s, ncol = s, byrow = TRUE)
 
 f_dist <- array(c(f_dist_1, f_dist_2), dim = c(s, s, degree + 1))
 f_dist_pars <- array(c(f_dist_1_pars_1, f_dist_1_pars_2,
@@ -209,7 +207,7 @@ f_dist_pars <- array(c(f_dist_1_pars_1, f_dist_1_pars_2,
 
 Then, defining a `dsmm_parametric` object is done simply through the function 
 `parametric_dsmm()`:
-```r
+```{r, eval = FALSE}
[…]
