### 2024-07-21 8db0dfd
Finalizing changes for paper before sending to Vlad and Nicolas
