+++ paper.md
@@ -18,11 +18,9 @@ authors:
 affiliations:
   - name: Laboratory of Mathematics Raphaël Salem (LMRS), UMR CNRS 6085, University of Rouen Normandy, France
     index: 1
-  - name: Laboratoire de Mécanique, Modélisation & Procédés Propres (M2P2), M2P2 UMR 7340, Aix Marseille Université, CNRS, Centrale Méditerranée, France 
+  - name: Laboratoire de Mécanique, Modélisation & Procédés Propres (M2P2), UMR 7340, Aix Marseille Université, CNRS, Centrale Méditerranée, France 
     index: 2
-  - name: Univ Rouen Normandie, CNRS, LMRS UMR 6085, F-76000 Rouen, France
-    index: 3
-date: 20 July 2024
+date: 21 July 2024
 bibliography: paper.bib
 editor_options: 
   markdown: 
@@ -31,23 +29,23 @@ editor_options:
 
 # Summary
 
-Markov models are a common approach in the researcher's toolbox in order to deal with the modeling of many real-life phenomenons presented as a sequence under discrete time. However, this imposes the assumption that the sequence under inspection is homogeneous with respect to time and furthermore that the sojourn times follow the Geometric distribution. This is not always true in practice when modeling, for example, DNA sequences.
+Markov models are a common approach in the researcher's toolbox for aiding with the modeling of many real-life phenomena, often presented as a sequence of states in discrete time, known as a Markov chain. However, this assumes that the sequence under inspection is time-homogeneous and that the sojourn times always follow a geometric distribution. Notice that this is not always true in practice, for example when modeling DNA sequences.
 
-Drifting semi-Markov models (DSMM) formulate the combination of semi-Markov models, which allow arbitrary distributions for the sojourn times, together with drifting Markov models, which describe the non-homogeneity of a sequence through a smooth, known shape that is gradually evolving, expressed through a linear (polynomial) function. As a result, DSMM are best suited to capture the non-homogeneity that occurs from the intrinsic evolution of the system or from the interactions between the system and the environment. For a detailed introduction to semi-Markov models see @barbu_limnios. Drifting Markov models were first introduced in @drift_polynomial.
+Drifting semi-Markov models (DSMM) aim to be a discrete-time pairing of semi-Markov models with drifting Markov models. Semi-Markov models allow an arbitrary choice for the distribution of the sojourn times. The drifting Markov models describe the non-homogeneity of a sequence through a smooth, known shape that is gradually evolving, expressed through a polynomial function. As a result, DSMM are best suited to capture non-homogeneities which occurs from the intrinsic evolution of the system or from the interactions between the system and the environment. For a detailed introduction to semi-Markov models see @barbu_limnios. Drifting Markov models were first introduced in @drift_polynomial.
 
-`dsmmR` is an R package which allows the user to perform parametric and non-parametric estimation and simulation of drifting semi-Markov processes. The user can also define their own parametric and non-parametric DSMM specifications, allowing for a necessary degree of freedom when dealing with a research question. Furthermore, three different types of DSMM are considered. These three models differ in the way they characterize the drifting semi-Markov kernel. Specifically, the first model allows both the transition matrix and the sojourn time distribution to vary (i.e. to "drift"), the second model allows only the transition matrix to drift while the third model allows only the sojourn time distribution to drift.
[…]
