+++ cran-comments.md
@@ -9,13 +9,14 @@ No ERRORs, WARNINGs or NOTEs.
 
 # R CMD check results through `rhub::rhub_check(branch = 'master')`.
 
-There are no ERRORs or NOTEs, only one WARNING.
+There are no ERRORs or NOTEs, only WARNINGs.
    
    Warnings:
    (1) * `rchk` (input number 19):
+      The warning reads as:
       `No files were found with the provided path: check. No artifacts will be uploaded.`
-      I believe this is an internal warning on `rchk`s side.
-      See <https://github.com/actions/upload-artifact/issues/232>
+      I believe this is an internal bug on `rchk`s side;
+      see <https://github.com/actions/upload-artifact/issues/232>
 
 
 # version 1.0.2 --------------------------------------------------------------------
