### 2024-07-27 84af489
adding debugging information in the github actions.
