### 2024-07-27 a43e022
testing new yaml workflow for more debugging info
