### 2024-07-27 59f3284
testing new yaml workflow for more debugging info
