### 2024-07-27 678ff35
Revergint to previous yaml configuration
