+++ .github/workflows/rhub.yaml
@@ -93,4 +93,3 @@ jobs:
         with:
           job-config: ${{ matrix.config.job-config }}
           token: ${{ secrets.RHUB_TOKEN }}
-      - name: List installed files
