### 2024-07-27 cb3de3d
Finalize before JOSS submission
