### 2024-07-27 65476f2
fixing weird error from rhub::check
