+++ R/fitmodel.R
@@ -629,12 +629,12 @@
 #' # ===========================================================================
 #'
 #'
-#' # sim_seq_nonparametric <- simulate(obj_model_1, nsim = 10)
-#' # str(sim_seq_nonparametric)
+#' sim_seq_nonparametric <- simulate(obj_model_1, nsim = 10)
+#' str(sim_seq_nonparametric)
 #'
 #'
-#' # kernel_drift_parametric <- get_kernel(obj_fit_parametric, klim = 5)
-#' # str(kernel_drift_parametric)
+#' kernel_drift_parametric <- get_kernel(obj_fit_parametric, klim = 10)
+#' str(kernel_drift_parametric)
 #'
 fit_dsmm <- function(sequence,
                      degree,
