### 2024-07-27 b3ab8fa
resetting to previous config for rhub.yaml file
