+++ R/fitmodel.R
@@ -633,7 +633,7 @@
 #' str(sim_seq_nonparametric)
 #'
 #'
-#' kernel_drift_parametric <- get_kernel(obj = obj_fit_parametric, klim = 10)
+#' kernel_drift_parametric <- dsmmR::get_kernel(obj = obj_fit_parametric, klim = 10)
 #' str(kernel_drift_parametric)
 #'
 fit_dsmm <- function(sequence,
