+++ man/fit_dsmm.Rd
@@ -572,12 +572,12 @@ cat("Not-drifting parameters for passing from ",
 # ===========================================================================
 
 
-sim_seq_nonparametric <- simulate(object = obj_model_1, nsim = 10)
-str(sim_seq_nonparametric)
+# sim_seq_nonparametric <- simulate(object = obj_model_1, nsim = 10)
+# str(sim_seq_nonparametric)
 
 
-kernel_drift_parametric <- dsmmR::get_kernel(obj = obj_fit_parametric, klim = 10)
-str(kernel_drift_parametric)
+# kernel_drift_parametric <- get_kernel(obj = obj_fit_parametric, klim = 10)
+# str(kernel_drift_parametric)
 
 }
 \references{
