+++ R/nonparametric.R
@@ -368,11 +368,11 @@
 #' # Using methods for non-parametric objects.
 #' # ===========================================================================
 #'
-#' # kernel_parametric <- get_kernel(obj = obj_nonpar_model_3)
-#' # str(kernel_parametric)
+#' kernel_parametric <- get_kernel(obj_nonpar_model_3)
+#' str(kernel_parametric)
 #'
-#' # sim_seq_par <- simulate(object = obj_nonpar_model_3, nsim = 50)
-#' # str(sim_seq_par)
+#' sim_seq_par <- simulate(obj_nonpar_model_3, nsim = 50)
+#' str(sim_seq_par)
 nonparametric_dsmm <- function(model_size,
                                states,
                                initial_dist,
