### 2024-07-27 #5572 @Mavrogiannis-Ioannis
Dear @Bisaloo,

1) The proof-reading is done alongside with my co-authors, and small modifications were made for clarity.
    You can download it here:
    [https://github.com/Mavrogiannis-Ioannis/dsmmR/actions/runs/10126436083](https://github.com/Mavrogiannis-Ioannis/dsmmR/actions/runs/10126436083).
2) The tagged release of the software is at:
    [https://github.com/Mavrogiannis-Ioannis/dsmmR/releases/tag/v1.0.5](https://github.com/Mavrogiannis-Ioannis/dsmmR/releases/tag/v1.0.5)
3) The reviewed software is in Zenodo, through the doi:
    [https://zenodo.org/doi/10.5281/zenodo.13117522](https://zenodo.org/doi/10.5281/zenodo.13117522)

Please, let me know for anything else necessary -- thank you so much for your work so far!

Respectfully,
Ioannis
