+++ README.md
@@ -5,15 +5,11 @@ README
 <!-- Use `build_readme()` to edit the .md file from the .rmd file -->
 <!-- badges: start -->
 
-[![CRAN
-status](https://www.r-pkg.org/badges/version/dsmmR)](https://CRAN.R-project.org/package=dsmmR)
+[![CRAN status](https://www.r-pkg.org/badges/version/dsmmR)](https://CRAN.R-project.org/package=dsmmR)
 [![R-CMD-check](https://github.com/Mavrogiannis-Ioannis/dsmmR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/Mavrogiannis-Ioannis/dsmmR/actions/workflows/R-CMD-check.yaml)
-[![CRAN
-logs](https://cranlogs.r-pkg.org/badges/dsmmR)](https://CRAN.R-project.org/package=dsmmR)
-[![Codecov test
-coverage](https://codecov.io/gh/Mavrogiannis-Ioannis/dsmmR/branch/master/graph/badge.svg)](https://app.codecov.io/gh/Mavrogiannis-Ioannis/dsmmR?branch=master)
-[![Contributor
-Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](code_of_conduct.md)
+[![CRAN logs](https://cranlogs.r-pkg.org/badges/dsmmR)](https://CRAN.R-project.org/package=dsmmR)
+[![Codecov test coverage](https://codecov.io/gh/Mavrogiannis-Ioannis/dsmmR/branch/master/graph/badge.svg)](https://app.codecov.io/gh/Mavrogiannis-Ioannis/dsmmR?branch=master)
+[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](https://usethis.r-lib.org/CODE_OF_CONDUCT.html)
 <!-- badges: end -->
 
 # Developer Version
@@ -58,14 +54,13 @@ For a detailed introduction in Drifting semi-Markov models consider the
 documentation through `?dsmmR`.
 
 For an extensive description of this approach, consider visiting the
-complete documentation of the package on the [official CRAN
-page](https://CRAN.R-project.org/package=dsmmR).
+complete documentation of the package on the [official CRAN page](https://CRAN.R-project.org/package=dsmmR).
 
 ### Estimation
 
 The easiest way to use **dsmmR** is through the main function
 `dsmm_fit()` in the non-parametric case. This function can estimate a
-Drifting semi-Markov model from a sequence of states (i.e. a character
+Drifting semi-Markov model from a sequence of states (i.e. a character
 vector in R). Example data is included in the package, defined in the
 DNA sequence `lambda`. Also some parameters need to be specified before
 using `dsmm_fit()`, most notably the polynomial *degree* and the model
@@ -301,26 +296,20 @@ c NA 0.50 NA
 
 ## Further reading
 
-Regarding semi-Markov models, the book [Semi-Markov Chains and Hidden
-Semi-Markov Models toward
-Applications](https://doi.org/10.1007/978-0-387-73173-5) gives a good
+Regarding semi-Markov models, the book [Semi-Markov Chains and Hidden Semi-Markov Models toward Applications](https://doi.org/10.1007/978-0-387-73173-5) gives a good
 overview of the topic and also combines the flexibility of the
 semi-Markov chain with the known advantages of hidden semi-markov
 models.
 
 If you are not familiar with Drifting Markov models, they were first
-introduced in [Drifting Markov models with Polynomial Drift and
-Applications to DNA Sequences](https://doi.org/10.2202/1544-6115.1326),
-while a comprehensive overview is provided in [Reliability and Survival
-Analysis for Drifting Markov Models: Modeling and
[…]
