+++ DESCRIPTION
@@ -2,7 +2,7 @@ Type: Package
 Package: dsmmR
 Title: Estimation and Simulation of Drifting Semi-Markov Models
 Version: 1.0.5
-Date: 2024-28-7
+Date: 2024-28-07
 Authors@R: c(
     person("Vlad Stefan", "Barbu",
            role = c("aut"),
@@ -13,7 +13,7 @@ Authors@R: c(
         comment = c(ORCID = "0000-0002-2948-0648")),
     person("Nicolas", "Vergne",
            role = c("aut"))
-           ) 
+           )
 Description: Performs parametric and non-parametric estimation and simulation of 
     drifting semi-Markov processes. The definition of parametric and non-parametric
     model specifications is also possible. Furthermore, three different types of
@@ -47,9 +47,5 @@ Suggests:
 Encoding: UTF-8
 RoxygenNote: 7.3.1
 NeedsCompilation: yes
-Author: Vlad Stefan Barbu [aut] (<https://orcid.org/0000-0002-0840-016X>),
-  Ioannis Mavrogiannis [aut, cre],
-  Nicolas Vergne [aut]
-Maintainer: Ioannis Mavrogiannis <mavrogiannis.ioa@gmail.com>
 Depends: R (>= 3.5.0)
 Config/testthat/edition: 3
