### 2024-07-29 #5572 @crvernon
:wave: @Mavrogiannis-Ioannis - I just need you to address the following few things before I move to accept this for publication:

- [ ] The license used in your GitHub repository does not match the one declared in your Zenodo archive.  Please edit the metadata of your Zenodo archive to include the correct license.  No need to conduct another release.

In the paper:

- [ ] Ensure that all items needing to maintain capitalization formatting do so in the references.  For example, in LINE 94 the "m" in "markov" should be capitalized.  Please check all of your references to ensure the correct capitalization has been maintained where needed.  You can do this in the bib file using curly brackets around the characters you wish to maintain formatting for.
