### 2024-07-30 2efdde6
Making JOSS paper references Capital
