+++ paper.bib
@@ -9,14 +9,14 @@
 
 
 @Article{barbu_semi_reliability,
-  title={smmR: Semi-{Markov} Models, {Markov} Models and Reliability},
+  title={smmR: {Semi-{Markov} Models, {Markov} Models and Reliability}},
   author={Barbu, Vlad Stefan and Lecocq, Florian and Lothode, Corentin}
 }
 
 
 
 @Manual{smmR_package,
-    title = {smmR: Simulation, Estimation and Reliability of Semi-{Markov} Models},
+    title = {smmR: {Simulation, Estimation and Reliability of Semi-{Markov} Models}},
     author = {Vlad Stefan Barbu and Florian Lecocq and Corentin Lothode and Nicolas Vergne},
     year = {2021},
     note = {R package version 1.0.3},
@@ -26,7 +26,7 @@
 
 
 @Article{drift_polynomial,
-  title = {Drifting {Markov} Models with Polynomial Drift and Applications to DNA Sequences},
+  title = {{Drifting {Markov} Models with Polynomial Drift and Applications to DNA Sequences}},
   author = {Nicolas Vergne},
   volume = {7},
   number = {1},
@@ -38,7 +38,7 @@
 
 
 @Article{drift_reliability,
-  title={Reliability and survival analysis for drifting {Markov} models: modeling and estimation},
+  title={{Reliability and survival analysis for drifting {Markov} models: modeling and estimation}},
   author={Barbu, Vlad Stefan and Vergne, Nicolas},
   journal={Methodology and Computing in Applied Probability},
   volume={21},
@@ -83,7 +83,7 @@
 
 
 @Manual{HMM_package,
-  title = {HMM: Hidden Markov Models},
+  title = {HMM: {Hidden Markov Models}},
   author = {Lin Himmelmann},
   year = {2022},
   note = {R package version 1.0.1},
@@ -92,7 +92,7 @@
 }
 
 @Manual{HiddenMarkov_package,
-    title = {HiddenMarkov: Hidden Markov Models},
+    title = {{HiddenMarkov: Hidden Markov Models}},
     author = {David Harte},
     year = {2021},
     note = {R package version 1.8-13},
@@ -103,7 +103,7 @@
   }
 
 @Article{mhsmm_package,
-    title = {Hidden Semi Markov Models for Multiple Observation Sequences: The {mhsmm} Package for {R}},
+    title = {{Hidden Semi Markov Models for Multiple Observation Sequences: The {mhsmm} Package for {R}}},
     author = {Jared O'Connell and Søren Højsgaard},
     journal = {Journal of Statistical Software},
     year = {2011},
@@ -113,3 +113,4 @@
     url = {http://www.jstatsoft.org/v39/i04/},
     doi = {10.18637/jss.v039.i04}
   }
+
