### 2024-07-30 dc74a17
JOSS paper capitalization
