+++ paper.bib
@@ -61,7 +61,7 @@
 
 
 @Article{drifting_learning,
-  title={{INVESTIGATING INDIVIDUAL PATHWAYS OF LEARNING: MODELLING THE DYNAMICS OF LEARNING WITH DRIFTING {Markov} MODEL}},
+  title={INVESTIGATING INDIVIDUAL PATHWAYS OF LEARNING: MODELLING THE DYNAMICS OF LEARNING WITH DRIFTING {Markov} MODEL},
   author={Komar, John and Vergne, Nicolas and Seifert, Ludovic},
   journal={Universit{\'e} Le Havre Normandie France},
   pages={33},
@@ -70,7 +70,7 @@
 
 
 @Article{SemiMarkov_package,
-    title = {{SemiMarkov}: An {R} Package for Parametric Estimation in Multi-State Semi-Markov Models},
+    title = {{SemiMarkov}: {An {R} Package for Parametric Estimation in Multi-State Semi-{Markov} Models}},
     author = {Agnieszka Kr\'{o}l and Philippe Saint-Pierre},
     journal = {Journal of Statistical Software},
     year = {2015},
@@ -103,7 +103,7 @@
   }
 
 @Article{mhsmm_package,
-    title = {{Hidden Semi Markov Models for Multiple Observation Sequences: The {mhsmm} Package for {R}}},
+    title = {Hidden {Semi Markov Models for Multiple Observation Sequences: The {mhsmm} Package for {R}}},
     author = {Jared O'Connell and Søren Højsgaard},
     journal = {Journal of Statistical Software},
     year = {2011},
