### 2024-07-30 1ac4f5f
JOSS paper references
