+++ paper.bib
@@ -13,8 +13,6 @@
   author={Barbu, Vlad Stefan and Lecocq, Florian and Lothode, Corentin}
 }
 
-
-
 @Manual{smmR_package,
     title = {smmR: {Simulation, Estimation and Reliability of Semi-{Markov} Models}},
     author = {Vlad Stefan Barbu and Florian Lecocq and Corentin Lothode and Nicolas Vergne},
