### 2024-07-31 #5572 @Mavrogiannis-Ioannis
Dear @crvernon, I have a small question:

In your journal, if an author is currently at institute X but research was conducted at institute Y,
which affiliation should be chosen?
Or, should a footnote be added clarifying the situation?
(e.g. Research was conducted mainly at Y, but author is currently at X)

The rest of what you have mentioned are already complete -- but this small detail for the affiliations still unsolved.
I would like to be sure about it before we proceed with the acceptance.

Thank you,
Ioannis
