### 2024-07-31 #5572 @crvernon
@Mavrogiannis-Ioannis, I don't think we have an official policy on this.  I usually use the affiliation that the author had when they conducted the research.  But you should use which one you feel comfortable with.

Once you are comfortable with how it is stated let me know and I'll proceed.
