### 2024-08-05 c3c98ff
Add affiliations at README file
