### 2024-08-05 #5572 @Mavrogiannis-Ioannis
Dear @crvernon , I am alright with the affiliations.

- I re-submitted the software tag, with the slight updated changes.
- It has the same name, v1.0.5.
- I also updated the affiliations for the zenodo repository.

How do we proceed?

Best,
Ioannis
